# This is a combination of 2 commits. # The first commit's message is:
Swipe to archive.

# This is the 2nd commit message:

Fixed cache cleaning vs. storing timestamp bug.

We were updating the list-last-updated timestamp even when there was a search
string specified. Which caused the server only to send filtered results. And
when we updated the timestamp we'd miss all the new event that didn't match the
search string.